Fix bug causing app to have a grey overlay screw up the whole UI and refuse touches.
As the main activity listenes to changes in preference, we only want
to update the UI when there is change on SyncStatus preference changes.
The way we tried to filter SyncStatus related changes was wrong and
we ended up creating multiple dialogs that finally added too many
dialogs(though they were invisible) creating a greyed out overlay
and blocking the app.