Fix bug causing syncing to fail saying "internet Problem".
It was a problem with the query selection criteria used when generating the sync request json.
Because of the way category fetching also fetches their children, the query used is a JOIN query, and
selection strings should also consider that. Otherwise we will try and lookup a non-existing column
from the database resulting in exception.